requirements.txt
# Async and HTTP Libraries
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
multidict==6.0.5
yarl==1.9.4
httpcore==1.0.5
httpx==0.27.0

# Data and JSON Handling
attrs==23.2.0
beautifulsoup4==4.12.3
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
dataclasses-json==0.6.7
dirtyjson==1.0.8
pyarrow==17.0.0
PyYAML==6.0.1

# Data Analysis and Machine Learning
altair==5.3.0
joblib==1.4.2
numpy==1.26.4
pandas==2.2.2
networkx==3.3
nltk==3.8.1

# Visualization
pydeck==0.9.1
Pygments==2.18.0

# Web Framework and Middleware
blinker==1.8.2
Jinja2==3.1.4
MarkupSafe==2.1.5
SQLAlchemy==2.0.31

# Python Utilities and Extensions
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
distro==1.9.0
exceptiongroup==1.2.2
frozenlist==1.4.1
mypy-extensions==1.0.0
nest-asyncio==1.6.0
packaging==24.1
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
striprtf==0.0.26
tenacity==8.5.0
tornado==6.4.1
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wrapt==1.16.0

# Streamlit and related libraries
streamlit==1.36.0

# Additional Tools and Libraries
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
Markdown-it-py==3.0.0
mdurl==0.1.2
pydantic==2.8.2
pydantic_core==2.20.1
pypdf==4.3.1
python-dotenv==1.0.1
rich==13.7.1
rpds-py==0.19.0
toolz==0.12.1

# Llama Libraries
llama-cloud==0.0.9
llama-index==0.10.57
llama-index-agent-openai==0.2.9
llama-index-cli==0.1.12
llama-index-core==0.10.57
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.2.5
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.27
llama-index-multi-modal-llms-openai==0.1.8
llama-index-program-openai==0.1.7
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.30
llama-index-readers-llama-parse==0.1.6
llama-parse==0.4.9
tiktoken==0.7.0

# OpenAI Libraries
openai==1.37.0
